Factor out a quickcheck-classes-base library that primitive can use in its test suite without causing a circular dependency
