+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Alternative.hs
