+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Applicative.hs
