+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Arrow.hs
