+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Bifoldable.hs
