+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Bifunctor.hs
