+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Bitraversable.hs
