+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Bits.hs
