+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Category.hs
