+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Contravariant.hs
