+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Enum.hs
