+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Eq.hs
