+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Foldable.hs
