+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Functor.hs
