+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Generic.hs
