+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Integral.hs
