+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Ix.hs
