+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Monad.hs
