+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/MonadFail.hs
