+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/MonadPlus.hs
