+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/MonadZip.hs
