+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Monoid.hs
