+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Num.hs
