+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Ord.hs
