+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Semigroup.hs
