+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Show.hs
