+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Traversable.hs
